+++ app/jobs/tracks/create_job.rb
@@ -1,11 +1,12 @@
 # frozen_string_literal: true
 
 class Tracks::CreateJob < ApplicationJob
-  queue_as :default
+  queue_as :tracks
 
-  def perform(user_id, start_at: nil, end_at: nil, cleaning_strategy: :replace)
+  def perform(user_id, start_at: nil, end_at: nil, mode: :daily)
     user = User.find(user_id)
-    tracks_created = Tracks::CreateFromPoints.new(user, start_at:, end_at:, cleaning_strategy:).call
+
+    tracks_created = Tracks::Generator.new(user, start_at:, end_at:, mode:).call
 
     create_success_notification(user, tracks_created)
   rescue StandardError => e