+++ spec/services/points_limit_exceeded_spec.rb
@@ -24,15 +24,20 @@ RSpec.describe PointsLimitExceeded do
 
       context 'when user points count is equal to the limit' do
         before do
-          allow(user.points).to receive(:count).and_return(10)
+          allow(user.tracked_points).to receive(:count).and_return(10)
         end
 
         it { is_expected.to be true }
+
+        it 'caches the result' do
+          expect(user.tracked_points).to receive(:count).once
+          2.times { described_class.new(user).call }
+        end
       end
 
       context 'when user points count exceeds the limit' do
         before do
-          allow(user.points).to receive(:count).and_return(11)
+          allow(user.tracked_points).to receive(:count).and_return(11)
         end
 
         it { is_expected.to be true }
@@ -40,7 +45,7 @@ RSpec.describe PointsLimitExceeded do
 
       context 'when user points count is below the limit' do
         before do
-          allow(user.points).to receive(:count).and_return(9)
+          allow(user.tracked_points).to receive(:count).and_return(9)
         end
 
         it { is_expected.to be false }